Rust Cargo linter: Support workspaces, any current directory
When using Cargo workspaces [1], there is a 'Cargo.toml' directory in a
top level directory, listing all the crates in the project. If we are
currently editing a single crate, 'cargo build' should execute in that
directory for that crate's separate `Cargo.toml`, otherwise the build
won't work.

This also assist in situations where the Vim current directory is
outside crate of the '.rs' file we are editing.

[1] https://doc.rust-lang.org/book/second-edition/ch14-03-cargo-workspaces.html